traffic_congestion/data_collection/pyproject.toml
[tool.poetry]
name = "data-collection"
version = "0.1.0"
description = "Using AI to support the local community problem."
authors = ["tqtensor - <victortang212@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<=3.19"
click = "^8.1.3"
ffmpy = "^0.3.0"
google-cloud-storage = "^2.6.0"
m3u8 = "^3.3.0"
py = "^1.11.0"
pydub = "^0.25.1"
pytest = "^7.2.0"
requests = "^2.28.1"
tqdm = "^4.64.1"
python-dotenv = "^0.21.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
yapf = "^0.32.0"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
